refactor: move MaxComputeProtobufConverterCache.java to converter package
